[WIP] Add Kinvey auth support
Add the first pass at Kinvey auth.  The Kinvey auth flow requires that
we check to see if a user exists, if not we create the user with the
access token from MIC.  If the user already exists we just log them in.
This gives back the Kinvey access token.